requirements.txt
textual-dev
python-dotenv
websockets